Viewport: replace unwieldy barycentric calculations
Replace them with a cross-product function that is much faster and
doesn't cause "tears" when a tile falls on the centerline between the
two triangles the frustum was split up in.

+++ viewport.h
@@ -20,7 +20,7 @@ void viewport_screen_to_world (double sx, double sy, double *wx, double *wy);
 void viewport_world_to_screen (double wx, double wy, double *sx, double *sy);
 void viewport_get_frustum (double **wx, double **wy);
 void viewport_get_bbox (double **bx, double **by);
-bool point_inside_frustum (double x, double y);
+bool point_inside_frustum (float x, float y);
 
 unsigned int viewport_get_ht (void);
 unsigned int viewport_get_wd (void);
